feat(UIKIT-599,ui,Notification): исправления по замечаниям из PR
