+++ packages/components/src/Notification/NotificationTemplate/constants.tsx
@@ -1,20 +1,15 @@
 import { ToastProps } from 'react-toastify/dist/types';
 
-import { NotificationTemplate } from '../NotificationTemplate';
-import { NotificationProps, Variant } from '../types';
+import { NotificationVariantTypes } from '../constants';
 import { notificationTemplatePropsCreator } from '../utils';
+import { NotificationProps, Variant } from '../types';
+
+import { NotificationTemplate } from './NotificationTemplate';
 
 type NotifyAction = (
   options: NotificationProps,
   notifyProps: ToastProps,
 ) => JSX.Element;
-
-export enum NotificationVariantTypes {
-  info = 'info',
-  success = 'success',
-  warning = 'warning',
-  error = 'error',
-}
 
 export const NOTIFICATION_VARIANT: Record<Variant, NotifyAction> = {
   info: (options, notifyContent) => (
