--- packages/components/src/Notification/constants/defaultNotificationProps.ts
@@ -1,3 +1,0 @@
-export const DEFAULT_NOTIFICATION_PROPS = {
-  filled: true,
-};
