--- packages/components/src/Notification/constants/index.ts
@@ -1,7 +1,0 @@
-export * from './positions';
-
-export * from './notificationVariant';
-
-export * from './defaultNotificationProps';
-
-export * from './notifyClassName';
