--- packages/components/src/Notification/constants/notifyClassName.ts
@@ -1,1 +1,0 @@
-export const NOTIFY_CLASSNAME = 'notify';
