--- packages/components/src/Notification/constants/positions.ts
@@ -1,3 +1,0 @@
-import { toast } from 'react-toastify';
-
-export const NOTIFY_POSITIONS = toast.POSITION;
